limit rotation to [-pi,pi>

+++ src/kalman/kalmanObject.cpp
@@ -135,7 +135,7 @@ namespace rtt {
     double kalmanObject::limitRotation(double rotation) const{
         double constRot=fmod(rotation+M_PI, 2*M_PI)-M_PI;
         if (constRot<-M_PI||constRot>=M_PI){
-            return -M_PI+std::numeric_limits<double>::epsilon();
+            return -M_PI+std::numeric_limits<float>::epsilon();
         }
         return constRot;
     }
